Indicate perk picks which would trigger Perks Lottery
If Perks Lottery has been taken (or for the Perks Lottery perk choice itself) and taking a perk would not cause the other perk options to be removed, then a clover badge will be displayed.

Note this changes the results of `PerkInfoProvider.provide()` to be data objects containing the `perk_id`, `perk`, `gambled`, and `willBeRemoved` properties, instead of just the perk objects.

This does not update the trigger chance with taken lottery perks, as a player may want to know if they should skip early perk choices and backtrack once lottery has been taken to get more value.